src/requirements.txt
Django==2.2.10
uWSGI==2.0.18
flake8==3.7.9
django-environ==0.4.5
django-debug-toolbar==2.2
django-anymail==7.0.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
Pillow==7.0.0
drf-yasg==1.17.1
reportlab==3.5.34
requests==2.23.0
python-dotenv==0.12.0
psycopg2-binary==2.8.4